refactor LimitExtractor
