+++ sharding-core/src/main/resources/META-INF/parsing-rule-definition/common/extractor-rule-definition.xml
@@ -14,6 +14,5 @@
     <extractor-rule id="fromWhere" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.FromWhereExtractor" />
     <extractor-rule id="groupBy" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.dql.GroupByExtractor" />
     <extractor-rule id="orderBy" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.dql.OrderByExtractor" />
-    <extractor-rule id="limit" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.dql.LimitExtractor" />
-    <extractor-rule id="subQueryCondition" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.SubQueryConditionExtractor" />
+    <extractor-rule id="subQueryCondition" extractor-class="io.shardingsphere.core.parsing.antlr.extractor.impl.dql.SubQueryConditionExtractor" />
 </extractor-rule-definition>
